New `geometry` module.
The mesh data structures defined in this module allows simple mesh manipulations and will solve the code repetition in the `med` and `abaqus` modules. This way, all the computational geometry functionalities will reside in the `geometry` module, while the `med` and `abaqus` modules will only deal with what their names refer to.