Removing unnecessary features from the `med` module.
The previous commit (3842041) allowed me to delete the non-MED related functions from the `med` module. The similar functions could have been removed from the `abaqus` module as well, but that module will anyway need a complete overhaul. Notice that the script now saves data to a numpy file, which is smaller and more portable than an Abaqus .inp file.