Make cross-references work in the documentation.
The `:noindex:` option was used to suppress the `duplicate object description` warnings in Sphinx. As the `automodule` command fetches the same objects as the module level documentation, this makes sense.
These types of warnings are caused by `:toctree: classes/` and `:toctree: functions/` in the module level documentation. The reason I need them is because this way the functions and classes appear in the TOC.
The warnings do not seem to indicate a visible problem.