Documentation: reference to the MED module brings to its documentation.
The target is set to an .rst file but the reference is put to a .md file.
This works well due to the recommonmark Sphinx extension that makes the .md file see the targets in .rst files.
However, when viewed in a Markdown viewer (e.g. on GitHub), the target will not be recognised. This is not a big problem as the document in which the reference was added is supposed to be viewed as part of the whole documentation (as opposed to e.g. the README file).