Do not support junction-based neighborhood search.
That method was quite fragile, therefore the possible speed-up does not justify its use. The emphasis must be on robustness.

The neighborhood search function now returns a tuple instead of a list. The tuple is created from a generator, as described on https://stackoverflow.com/a/16940351/4892892
Reasons for returning a tuple instead of a list:
- It emphasizes the fact that the output is not intended to be modified (immutable)
- The returned tuple is used for indexing in the `skeleton2regions` function of the `cad` module. If I use a list for indexing, I get the following warning from numpy: "FutureWarning: Using a non-tuple sequence for multidimensional indexing is deprecated; use `arr[tuple(seq)]` instead of `arr[seq]`."