Multiple grains touching each other are allowed to be embedded.
Until now, the embedded grains were identified by being incident to an isolated cycle.
However, this efficient strategy only worked for single embedded grains.
Once a cluster of grains are embedded, the isolated cycles are not indicative.
We use an elegant brute-force algorithm to detect these cases too. But it is very slow!